Move fluid transport mechanics into the pipe domain
