+++ common/src/client/java/com/logistics/fluid/render/FluidPipeBlockEntityRenderer.java
@@ -9,8 +9,8 @@ import com.logistics.core.lib.resource.ResourceId;
 import com.logistics.fluid.block.FluidConnection;
 import com.logistics.fluid.block.FluidPipeBlock;
 import com.logistics.fluid.block.entity.FluidPipeBlockEntity;
-import com.logistics.fluid.pipe.FluidColumnGeometry;
-import com.logistics.fluid.pipe.FluidPipeModules;
+import com.logistics.pipe.fluid.FluidColumnGeometry;
+import com.logistics.pipe.fluid.FluidPipeModules;
 import com.logistics.pipe.render.model.PipeGeometry;
 import com.logistics.pipe.render.model.PipeModelResolver;
 import com.mojang.blaze3d.vertex.PoseStack;
