+++ common/src/main/java/com/logistics/fluid/block/entity/FluidPipeBlockEntity.java
@@ -22,12 +22,12 @@ import com.logistics.pipe.fluid.FluidUnits;
 import com.logistics.fluid.block.FluidConnection;
 import com.logistics.fluid.block.FluidPipeBlock;
 import com.logistics.fluid.block.FluidPipeKind;
-import com.logistics.fluid.pipe.FluidExtraction;
-import com.logistics.fluid.pipe.FluidPipe;
-import com.logistics.fluid.pipe.FluidPipeModules;
-import com.logistics.fluid.pipe.FluidProvider;
-import com.logistics.fluid.pipe.FluidSplit;
-import com.logistics.fluid.pipe.TravelingFluid;
+import com.logistics.pipe.fluid.FluidExtraction;
+import com.logistics.pipe.fluid.FluidPipe;
+import com.logistics.pipe.fluid.FluidPipeModules;
+import com.logistics.pipe.fluid.FluidProvider;
+import com.logistics.pipe.fluid.FluidSplit;
+import com.logistics.pipe.fluid.TravelingFluid;
 import com.mojang.serialization.DataResult;
 import java.util.ArrayList;
 import java.util.List;
