+++ common/src/main/java/com/logistics/pipe/fluid/FluidPipeModules.java
@@ -1,4 +1,4 @@
-package com.logistics.fluid.pipe;
+package com.logistics.pipe.fluid;
 
 import com.logistics.core.lib.pipe.CoreDecoration;
 import com.logistics.core.lib.pipe.IModuleHost;
